Redesigned the Button class. Inheritance was used, Keyboard Button and Menu Button classes were added.
